requirements.txt
# ==============================================================================
# RAP NOVA ESCOLA - Dependências do Projeto
# ==============================================================================
# Este ficheiro lista todas as bibliotecas Python necessárias para o projeto
# 
# Como instalar:
# 1. Abrir terminal na pasta do projeto
# 2. Executar: pip install -r requirements.txt
# 
# Nota: Recomenda-se usar um ambiente virtual (venv)
# ==============================================================================

# Base de Dados PostgreSQL
# -------------------------
# psycopg2-binary é o driver para conectar Python ao PostgreSQL
# Permite executar queries, criar tabelas, inserir dados, etc
psycopg2-binary==2.9.9

# Gestão de Variáveis de Ambiente
# ---------------------------------
# python-dotenv lê o ficheiro .env e carrega as variáveis
# Útil para guardar passwords e configurações sensíveis
python-dotenv==1.0.0

# Validação de Dados
# -------------------
# pydantic ajuda a validar dados (emails, números, datas)
# Garante que os dados inseridos estão no formato correto
pydantic==2.5.0

# Datas e Horas
# --------------
# python-dateutil facilita trabalhar com datas e horários
# Útil para calcular disponibilidades e agendamentos
python-dateutil==2.8.2

# Tabelas Bonitas no Terminal (Opcional mas útil)
# ------------------------------------------------
# tabulate permite mostrar dados em formato de tabela no terminal
# Torna a visualização de listas e relatórios mais legível
tabulate==0.9.0

# Testes Automáticos
# -------------------
# pytest é usado para escrever e executar testes automáticos
# Garante que o código funciona como esperado
pytest==7.4.3
requests==2.31.0

# Framework para API Web
# ----------------------
# FastAPI para criar os endpoints da API
fastapi==0.104.1
# Servidor para correr a aplicação FastAPI
uvicorn[standard]==0.24.0

# Autenticação (validar JWT do Supabase)
# ---------------------------------------
PyJWT==2.8.0


# ORM (Object-Relational Mapping) - OPCIONAL para versões futuras
# -----------------------------------------------------------------
# SQLAlchemy facilita trabalhar com a BD usando objetos Python
# Comentado por agora, vamos usar SQL puro inicialmente
# sqlalchemy==2.0.23

# Logging Avançado - OPCIONAL para versões futuras
# --------------------------------------------------
# loguru torna os logs mais bonitos e fáceis de ler
# Útil para debugging e monitorização
# loguru==0.7.2
supabase>=2.12.0
httpx==0.24.1
